create layout and container for pages
