+++ package.json
@@ -18,7 +18,8 @@
     "react-scripts": "2.0.5",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.5",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "sweetalert2": "^7.28.11"
   },
   "scripts": {
     "start": "react-scripts start",
