+++ src/actions/authActions.js
@@ -1,6 +1,7 @@
 
 import axios from 'axios';
 import queryString from 'query-string';
+import swal from 'sweetalert2';
 import * as types from './actionTypes';
 
 
@@ -8,6 +9,11 @@ export const forgotPassword = (userData) => async (dispatch) => {
   try {
     const response = await axios.post('api/forgotPassword', userData);
     dispatch({ type: types.SUCCESS_MSG, payload: response.data.message });
+    swal(
+      response.data.message,
+      'Click the link in the email to reset your password',
+      'success',
+    );
   } catch (error) {
     dispatch({ type: types.RESET_ERROR, payload: error.response.data.message });
   }
@@ -17,6 +23,12 @@ export const resetPassword = (userData, history) => async (dispatch) => {
     const { token } = queryString.parse(history.location.search);
     const response = await axios.post(`api/resetPassword?token=${token.trim()}`, userData);
     dispatch({ type: types.SUCCESS_MSG, payload: response.data.message });
+    swal({
+      title: response.data.message,
+      type: 'success',
+      html: '<a href="/login">Login </a>',
+      showConfirmButton: false,
+    });
   } catch (error) {
     dispatch({ type: types.RESET_ERROR, payload: error.response.data.message });
   }
