+++ src/containers/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
-import ForgotPassword from '../views/ForgotPassword';
-import ResetPassword from '../views/ResetPassword';
+import ForgotPasswordPage from '../views/ForgotPasswordPage';
+import ResetPasswordPage from '../views/ResetPasswordPage';
 import Auth from '../views/Auth';
 import Hero from '../views/Hero';
 import Profile from '../views/Profile';
@@ -21,8 +21,8 @@ const App = () => (
     <div>
       <Route exact path="/" component={Hero} />
       <Route path="/auth" component={Auth} />
-      <Route path="/forgotpassword" component={ForgotPassword} />
-      <Route path="/resetpassword" component={ResetPassword} />
+      <Route path="/forgotpassword" component={ForgotPasswordPage} />
+      <Route path="/resetpassword" component={ResetPasswordPage} />
       <Route path="/user" component={authenticate(User)} />
     </div>
   </BrowserRouter>
